Add files via upload
The Electronic Store programme is a based on use of ADT's and principles of OOP. The programme has an electronic store which has products, registers customers and contains different operations like selling the products, keeping track of the stock, registering new customers, keeping track of purchase history, sorting the products ,etc.

The Napster programme simulates an on-line music centre called the MusicExchangeCenter
where Users log in and download Songs from other users’ computers. It is a similar idea to the operation
of the original Napster program that operated many years ago where users shared songs. The programme deals with hash maps and lists.

The programme GUI uses Graphical interface to make a window which lists DVD's. The user can add or delete DVD's into the window.